+++ ontimize/components/form/toolbar/o-form-toolbar.component.ts
@@ -2,13 +2,13 @@ import { Component, OnInit, OnDestroy, Inject, Injector, forwardRef, ElementRef,
 import { CommonModule } from '@angular/common';
 import { Subscription } from 'rxjs';
 
-import { OFormComponent } from './o-form.component';
-import { InputConverter } from '../../decorators';
-import { Util } from '../../util/util';
-import { PermissionsUtils } from '../../util/permissions';
-import { DialogService, NavigationService, OPermissions, SnackBarService } from '../../services';
-import { OSharedModule } from '../../shared';
-import { OFormNavigationComponent } from './navigation/o-form-navigation.component';
+import { InputConverter } from '../../../decorators';
+import { Util } from '../../../util/util';
+import { PermissionsUtils } from '../../../util/permissions';
+import { DialogService, NavigationService, OPermissions, SnackBarService } from '../../../services';
+import { OSharedModule } from '../../../shared';
+import { OFormNavigationComponent } from '../navigation/o-form-navigation.component';
+import { OFormComponent } from '../o-form.component';
 
 export const DEFAULT_INPUTS_O_FORM_TOOLBAR = [
   'labelHeader: label-header',